added InsightParser.cs
