+++ modules/back-end/src/Application/Services/IInsightService.cs
@@ -5,8 +5,6 @@ namespace Application.Services;
 
 public interface IInsightService
 {
-    bool TryParse(string json, out object insight);
-
     Task AddManyAsync(object[] insights);
 
     Task<ICollection<Insight>> GetInsightsAsync(Guid envId, InsightFilter filter);
