+++ modules/back-end/src/Infrastructure/MQ/InsightMessageHandler.cs
@@ -3,15 +3,15 @@ using Domain.Messages;
 
 namespace Infrastructure.MQ;
 
-public class InsightMessageHandler(IInsightService insightService, InsightsTracker insightsTracker) : IMessageHandler
+public class InsightMessageHandler(InsightsTracker insightsTracker) : IMessageHandler
 {
     public string Topic => Topics.Insights;
 
     public async Task HandleAsync(string message)
     {
-        if (insightService.TryParse(message, out var insight))
+        if (InsightParser.TryParse(message, out var insight))
         {
-            await insightsTracker.RecordAsync(insight);
+            await insightsTracker.RecordAsync(insight!);
         }
     }
 }
