+++ modules/back-end/src/Infrastructure/Services/MongoDb/InsightService.cs
@@ -1,8 +1,6 @@
-using System.Text.Json.Nodes;
 using Application.FeatureFlags;
 using Application.Insights;
 using Domain.Experiments;
-using MongoDB.Bson;
 using MongoDB.Driver;
 
 namespace Infrastructure.Services.MongoDb;
@@ -15,60 +13,15 @@ public class InsightService(MongoDbClient mongoDb) : IInsightService
         IsOrdered = false
     };
 
-    public bool TryParse(string json, out object? insight)
-    {
-        try
-        {
-            insight = Parse();
-        }
-        catch
-        {
-            insight = null;
-        }
-
-        return insight != null;
-
-        object Parse()
-        {
-            var jsonNode = JsonNode.Parse(json)!.AsObject();
-
-            jsonNode["_id"] = jsonNode["uuid"]!.GetValue<string>();
-            jsonNode.Remove("uuid");
-
-            jsonNode["properties"] = JsonNode.Parse(jsonNode["properties"]!.GetValue<string>());
-
-            var timestampInMilliseconds = jsonNode["timestamp"]!.GetValue<long>() / 1000;
-            var timestamp = DateTimeOffset.FromUnixTimeMilliseconds(timestampInMilliseconds).UtcDateTime;
-            jsonNode["timestamp"] = timestamp;
-
-            var bsonDocument = BsonDocument.Parse(jsonNode.ToJsonString());
-            bsonDocument["timestamp"] = timestamp;
-
-            return bsonDocument;
-        }
-    }
-
     public async Task AddManyAsync(object[] insights)
     {
-        var documents = insights.Cast<BsonDocument>().ToArray();
-
-        var exposures = documents
-            .Select(TryBuildExposure)
-            .Where(x => x != null)
-            .Cast<ExperimentExposureEvent>()
-            .ToArray();
-
+        var exposures = insights.OfType<ExperimentExposureEvent>().ToArray();
         if (exposures.Length > 0)
         {
             await mongoDb.CollectionOf<ExperimentExposureEvent>().InsertManyAsync(exposures, InsertOptions);
         }
 
-        var metrics = documents
-            .Select(TryBuildMetric)
-            .Where(x => x != null)
-            .Cast<ExperimentMetricEvent>()
-            .ToArray();
-
+        var metrics = insights.OfType<ExperimentMetricEvent>().ToArray();
         if (metrics.Length > 0)
         {
             await mongoDb.CollectionOf<ExperimentMetricEvent>().InsertManyAsync(metrics, InsertOptions);
@@ -107,110 +60,6 @@ public class InsightService(MongoDbClient mongoDb) : IInsightService
                     .ToArray()
             })
             .ToArray();
-    }
-    
-    private static ExperimentExposureEvent? TryBuildExposure(BsonDocument doc)
-    {
-        if (doc.GetValue("event", string.Empty).AsString != "FlagValue" ||
-            !Guid.TryParse(doc.GetValue("env_id", string.Empty).AsString, out var envId))
-        {
-            return null;
-        }
-
-        var properties = doc.GetValue("properties", new BsonDocument()).AsBsonDocument;
-        var flagKey = GetString(properties, "featureFlagKey");
-        var userKey = GetString(properties, "userKeyId");
-        var variationId = GetString(properties, "variationId");
-
-        if (string.IsNullOrWhiteSpace(flagKey) ||
-            string.IsNullOrWhiteSpace(userKey) ||
-            string.IsNullOrWhiteSpace(variationId))
-        {
-            return null;
-        }
-
-        return new ExperimentExposureEvent
-        {
-            Id = GetId(doc),
-            EnvId = envId,
-            FlagKey = flagKey,
-            UserKey = userKey,
-            VariationId = variationId,
-            VariationValue = GetString(properties, "variationValue"),
-            ExposedAt = doc.GetValue("timestamp").ToUniversalTime(),
-            Properties = properties.ToJson(),
-            CreatedAt = DateTime.UtcNow
-        };
-    }
-
-    private static ExperimentMetricEvent? TryBuildMetric(BsonDocument doc)
-    {
-        var eventType = doc.GetValue("event", string.Empty).AsString;
-        if (eventType == "FlagValue" ||
-            !Guid.TryParse(doc.GetValue("env_id", string.Empty).AsString, out var envId))
-        {
-            return null;
-        }
-
-        var properties = doc.GetValue("properties", new BsonDocument()).AsBsonDocument;
-        var userKey = GetString(properties, "userKeyId") ?? GetNestedString(properties, "user", "keyId");
-        var eventName = GetString(properties, "eventName") ?? doc.GetValue("distinct_id", string.Empty).AsString;
-
-        if (string.IsNullOrWhiteSpace(userKey) || string.IsNullOrWhiteSpace(eventName))
-        {
-            return null;
-        }
-
-        return new ExperimentMetricEvent
-        {
-            Id = GetId(doc),
-            EnvId = envId,
-            UserKey = userKey,
-            EventName = eventName,
-            EventType = eventType,
-            NumericValue = GetNumericValue(properties),
-            OccurredAt = doc.GetValue("timestamp").ToUniversalTime(),
-            Properties = properties.ToJson(),
-            CreatedAt = DateTime.UtcNow
-        };
-    }
-
-    private static Guid GetId(BsonDocument doc)
-    {
-        var value = doc.GetValue("_id", BsonNull.Value);
-        return value.BsonType switch
-        {
-            BsonType.Binary => value.AsBsonBinaryData.ToGuid(),
-            BsonType.String when Guid.TryParse(value.AsString, out var id) => id,
-            _ => Guid.NewGuid()
-        };
-    }
-
-    private static string? GetString(BsonDocument doc, string key)
-    {
-        if (!doc.TryGetValue(key, out var value) || value.IsBsonNull)
-        {
-            return null;
-        }
-
-        return value.IsString ? value.AsString : value.ToString();
-    }
-
-    private static string? GetNestedString(BsonDocument doc, string objectKey, string key)
-    {
-        return doc.TryGetValue(objectKey, out var nested) && nested.IsBsonDocument
-            ? GetString(nested.AsBsonDocument, key)
-            : null;
-    }
-
-    private static double GetNumericValue(BsonDocument properties)
-    {
-        if (properties.TryGetValue("numericValue", out var numericValue) && numericValue.IsNumeric)
-        {
-            return numericValue.ToDouble();
-        }
-
-        return 0;
     }
 
     private static DateTime Truncate(DateTime timestamp, string intervalType)
